ambiguous isnan

+++ miniapps/meshing/mesh-optimizer.cpp
@@ -120,7 +120,7 @@ double RelaxedNewtonSolver::ComputeScalingFactor(const Vector &x,
       x_gf.SetFromTrueVector();
 
       energy_out = nlf->GetGridFunctionEnergy(x_gf);
-      if (energy_out > 1.2*energy_in || isnan(energy_out) != 0)
+      if (energy_out > 1.2*energy_in || isnan((double)energy_out) != 0)
       {
          if (print_level >= 0)
          { cout << "Scale = " << scale << " Increasing energy." << endl; }
